Solve Database Hostname issue
during Drupal installation, the DB Hostname was mapped to its specific
IP forcing the User to get the IP of the DB Container

Docker has the ability to map the container name to your local hosts

The DB container name is printed right after all containers have
started. Alternatively, it can be found via `docker ps` in the NAME
column.

As a result, the db container name will be used when Drupal installer
asks for db hostname

Resolves: #14
Resolves: #20
Signed-off-by: Minca Daniel Andrei <dminca@marketstheworld.com>